requirements.txt
aniso8601==9.0.1
click==8.0.1
Flask==2.0.0
Flask-MySQL==1.5.2
Flask-MySQLdb==0.2.0
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.9
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mysqlclient==2.0.3
pkg-resources==0.0.0
pymongo==3.11.4
PyMySQL==1.0.2
python-bsonjs==0.2.0
python-dotenv==0.17.1
pytz==2021.1
six==1.16.0
validator.py==1.3.0
waitress==2.0.0
Werkzeug==2.0.1
